refactor: query 관련 로직을 persistence로 분리
