+++ src/main/java/roomescape/persistence/ThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.persistence;
 
 import java.util.Optional;
 import roomescape.domain.Theme;
