+++ src/main/java/roomescape/persistence/util/RepositoryExceptionTranslator.java
@@ -1,4 +1,4 @@
-package roomescape.repository.util;
+package roomescape.persistence.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
