+++ src/main/java/roomescape/service/ReservationService.java
@@ -9,10 +9,10 @@ import roomescape.domain.ReservationSlot;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.exception.EntityNotFoundException;
-import roomescape.repository.ReservationSlotRepository;
-import roomescape.repository.ReservationTimeRepository;
-import roomescape.repository.ThemeRepository;
-import roomescape.repository.dto.ReservationCondition;
+import roomescape.persistence.ReservationSlotRepository;
+import roomescape.persistence.ReservationTimeRepository;
+import roomescape.persistence.ThemeRepository;
+import roomescape.persistence.dto.ReservationCondition;
 import roomescape.service.command.ReservationChangeCommand;
 import roomescape.service.command.ReservationCommand;
 import roomescape.service.result.ReservationSlotResult;
