+++ src/main/java/roomescape/service/command/ReservationChangeCommand.java
@@ -1,7 +1,7 @@
 package roomescape.service.command;
 
 import java.time.LocalDate;
-import roomescape.repository.dto.ReservationCondition;
+import roomescape.persistence.dto.ReservationCondition;
 
 public record ReservationChangeCommand(
         LocalDate date,
