+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -18,8 +18,8 @@ import roomescape.domain.fixture.ReservationTimeFixture;
 import roomescape.domain.fixture.ThemeFixture;
 import roomescape.exception.DuplicateEntityException;
 import roomescape.exception.EntityNotFoundException;
-import roomescape.repository.ReservationSlotRepository;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.persistence.ReservationSlotRepository;
+import roomescape.persistence.ReservationTimeRepository;
 import roomescape.service.command.ReservationChangeCommand;
 import roomescape.service.command.ReservationCommand;
 import roomescape.service.fake.FakeReservationSlotRepository;
