+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import roomescape.exception.DuplicateEntityException;
-import roomescape.repository.ThemeRepository;
+import roomescape.persistence.ThemeRepository;
 import roomescape.service.command.ThemeRegisterCommand;
 import roomescape.service.fake.FakeThemeRepository;
 import roomescape.service.result.ThemeRegisterResult;
