+++ src/test/java/roomescape/service/fake/FakeReservationTimeRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.persistence.ReservationTimeRepository;
 
 public class FakeReservationTimeRepository implements ReservationTimeRepository {
